% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_statistics_helpers.R
\name{as_srvyr_result_df}
\alias{as_srvyr_result_df}
\title{Create a srvyr results data.frame which is automatically unpacked by srvyr}
\usage{
as_srvyr_result_df(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
An object with the \code{srvyr_result_df} added
}
\description{
srvyr uses data.frame columns wrapped by this function to know when to
automatically unpack the results for the user. When developing extensions
(see vignette \code{extending-srvyr}), use this function to wrap the result
in so that \code{summarize} knows to unpack them.
}
\keyword{internal}
