% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_survey_rep.r
\name{as_survey_rep}
\alias{as_survey_rep}
\alias{as_survey_rep_}
\title{Create a tbl_svy survey object using replicate weights}
\usage{
as_survey_rep(.data, variables = NULL, repweights = NULL, weights = NULL,
  type = c("BRR", "Fay", "JK1", "JKn", "bootstrap", "other"),
  combined_weights = TRUE, rho = NULL, bootstrap_average = NULL,
  scale = NULL, rscales = NULL, fpc = NULL, fpctype = c("fraction",
  "correction"), mse = getOption("survey.replicates.mse"))

as_survey_rep_(.data, variables = NULL, repweights = NULL, weights = NULL,
  type = c("BRR", "Fay", "JK1", "JKn", "bootstrap", "other"),
  combined_weights = TRUE, rho = NULL, bootstrap_average = NULL,
  scale = NULL, rscales = NULL, fpc = NULL, fpctype = c("fraction",
  "correction"), mse = getOption("survey.replicates.mse"))
}
\arguments{
\item{.data}{A data frame (which contains the variables specified below)}

\item{variables}{Variables to include in the design (default is all)}

\item{repweights}{Variables specifying the replication weight varibles}

\item{weights}{Variables specifying sampling weights}

\item{type}{Type of replication weights}

\item{combined_weights}{\code{TRUE} if the \code{repweights} already
include the sampling weights. This is usually the case.}

\item{rho}{Shrinkage factor fo rweights in Fay's method}

\item{bootstrap_average}{For \code{type = "bootstrap"}, if the bootstrap
weights have been averaged, gives the number of iterations averaged over.}

\item{scale, rscales}{Scaling constant for variance, see
\code{\link[survey]{svrepdesign}} for more information.}

\item{fpc, fpctype}{Finite population correction information}

\item{mse}{if \code{TRUE}, compute varainces based on sum of squares
around the point estimate, rather than the mean of the replicates}
}
\value{
An object of class \code{tbl_svy}
}
\description{
A wrapper around \code{\link[survey]{svrepdesign}}. All survey variables must be included
in the data.frame itself. Select variables by using bare column names, or convenience
functions described in \code{\link[dplyr]{select}}. \code{as_survey_rep_} is the
standard evaluation counterpart to \code{as_survey_rep}
}
\examples{
# Examples from ?survey::svrepdesign()
library(survey)
data(scd)
# use BRR replicate weights from Levy and Lemeshow
scd$rep1 <- 2 * c(1, 0, 1, 0, 1, 0)
scd$rep2 <- 2 * c(1, 0, 0, 1, 0, 1)
scd$rep3 <- 2 * c(0, 1, 1, 0, 0, 1)
scd$rep4 <- 2 * c(0, 1, 0, 1, 1, 0)

scdrep <- scd \%>\%
  as_survey_rep(type = "BRR", repweights = starts_with("rep"),
                combined_weights = FALSE)

svyratio(~alive, ~arrests, scdrep)

}

