% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_tbl.r
\name{collect}
\alias{collect}
\alias{compute}
\title{Force computation of a database query}
\description{
\code{collect} retrieves data from a database query (and when run
on a tbl_svy object adjusts weights accordingly). Use collect when
you want to run a function from the survey package on a srvyr db
backed object. \code{compute} stores results in a remote temporary
table.
}
