% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_survey.r
\name{design_survey}
\alias{design_survey}
\alias{design_survey_}
\title{Create a tbl_svy survey object using sampling design}
\usage{
design_survey(.data, ids = NULL, probs = NULL, strata = NULL,
  variables = NULL, fpc = NULL, nest = FALSE, check.strata = !nest,
  weights = NULL, pps = FALSE, variance = c("HT", "YG"))

design_survey_(.data, ids = NULL, probs = NULL, strata = NULL,
  variables = NULL, fpc = NULL, nest = FALSE, check.strata = !nest,
  weights = NULL, pps = FALSE, variance = c("HT", "YG"))
}
\arguments{
\item{.data}{A data frame (which contains the variables specified below)}

\item{ids}{Variables specifying cluster ids from largest level to smallest level
(or NULL / nothing if no clusters).}

\item{probs}{Variables specifying cluster sampling probabilities.}

\item{strata}{Variables specifying strata.}

\item{variables}{Variables specifying variables to be included in survey.
Defaults to all variables in .data}

\item{fpc}{Variables specifying a finite population correct see
\code{\link[survey]{svydesign}} for more details.}

\item{nest}{If \code{TRUE}, relabel cluster ids to enforce nesting within strata.}

\item{check.strata}{If \code{TRUE}, check that clusters are nested in strata.}

\item{weights}{Variables specifying weights (inverse of probability).}

\item{pps}{"brewer" to use Brewer's approximation for PPS sampling without replacement.
"overton" to use Overton's approximation. An object of class HR to use the Hartley-Rao
approximation. An object of class ppsmat to use the Horvitz-Thompson estimator.}

\item{variance}{r pps without replacement, use variance="YG" for the Yates-Grundy estimator
instead of the Horvitz-Thompson estimator}
}
\value{
An object of class \code{tbl_svy}
}
\description{
A wrapper around \code{\link[survey]{svydesign}}. All survey variables must be included
in the data.frame itself. Select variables by using bare column names, or convenience
functions described in \code{\link[dplyr]{select}}. \code{design_survey_} is the
standard evaluation counterpart to \code{design_survey}
}
\examples{
# Examples from ?survey::svydesign
library(survey)
data(api)

# stratified sample
dstrata <- apistrat \%>\%
  design_survey(strata = stype, weights = pw)

# one-stage cluster sample
dclus1 <- apiclus1 \%>\%
  design_survey(dnum, weights = pw, fpc = fpc)

# two-stage cluster sample: weights computed from population sizes.
dclus2 <- apiclus2 \%>\%
  mutate(fpc2 = as.vector(fpc2)) \%>\% # work around dplyr's dislike of attributes
  design_survey(c(dnum, snum), fpc = c(fpc1, fpc2))

## multistage sampling has no effect when fpc is not given, so
## these are equivalent.
dclus2wr <- apiclus2 \%>\%
  dplyr::mutate(weights = weights(dclus2)) \%>\%
  design_survey(c(dnum, snum), weights = weights)

dclus2wr2 <- apiclus2 \%>\%
  dplyr::mutate(weights = weights(dclus2)) \%>\%
  design_survey(c(dnum), weights = weights)

## syntax for stratified cluster sample
## (though the data weren't really sampled this way)
apistrat \%>\% design_survey(dnum, strata = stype, weights = pw,
                           nest = TRUE)

## PPS sampling without replacement
data(election)
dpps <- election_pps \%>\%
  design_survey(fpc = p, pps = "brewer")

## design_survey_ uses standard evaluation
strata_var <- "stype"
weights_var <- "pw"
dstrata2 <- apistrat \%>\%
  design_survey_(strata = strata_var, weights = weights_var)

}

