% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_twophase.r
\name{design_twophase}
\alias{design_twophase}
\alias{design_twophase_}
\title{Create a tbl_svy survey object using two phase design}
\usage{
design_twophase(.data, id, strata = NULL, probs = NULL, weights = NULL,
  variables = NULL, subset, method = c("full", "approx", "simple"))

design_twophase_(.data, ids = NULL, probs = NULL, strata = NULL,
  variables = NULL, fpc = NULL, nest = FALSE, check.strata = !nest,
  weights = NULL, pps = FALSE, variance = c("HT", "YG"))
}
\arguments{
\item{.data}{A data frame (which contains the variables specified below)}

\item{id}{list of two sets of variable names for sampling unit identifers}

\item{strata}{list of two sets of variable names (or \code{NULLs}) for stratum identifiers}

\item{probs}{list of two sets of variable names (or \code{NULLs}) for sampling probabilities}

\item{weights}{Only for method = "approx", list of two sets of variable names (or \code{NULLs})
for sampling weights}

\item{method}{"full" requires (much) more memory, but gives unbiased variance estimates for
general multistage designs at both phases. "simple" or "approx" use less memory, and is correect for
designs with simple random sampling at phase one and stratifed randoms sampling at phase two. See
\code{\link[survey]{twophase}} for more details.}

\item{fpc}{list of two sets of variables (or \code{NULLs} for finite population corrections}
}
\value{
An object of class \code{tbl_svy}
}
\description{
A wrapper around \code{\link[survey]{twophase}}. All survey variables must be included
in the data.frame itself. Select variables by using bare column names, or convenience
functions described in \code{\link[dplyr]{select}}. \code{design_twophase_} is the
standard evaluation counterpart to \code{design_twophase}
}
\examples{
# Examples from ?survey::twophase
}

