% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manip.r
\name{dplyr_single}
\alias{dplyr_single}
\alias{filter}
\alias{filter_}
\alias{funs}
\alias{funs_}
\alias{mutate}
\alias{mutate_}
\alias{mutate_each}
\alias{mutate_each_}
\alias{rename}
\alias{rename_}
\alias{select}
\alias{select_}
\alias{summarise_each}
\alias{summarise_each_}
\alias{summarize_each}
\alias{summarize_each_}
\alias{transmute}
\alias{transmute_}
\title{single table verbs from dplyr}
\description{
These are data manipulation functions designed to work on \code{tbl_svy} objects.
}
\details{
See \code{\link[dplyr]{mutate}} for more details on \code{mutate} and \code{transmute}.

See \code{\link[dplyr]{select}} for more details on \code{select} and \code{rename}.

See \code{\link[dplyr]{filter}} for more details on \code{filter}.

\code{arrange} is not implemented for \code{tbl_svy} objects.
}

