% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.r
\name{groups}
\alias{groups}
\alias{ungroup}
\title{Get/set the grouping variables for tbl.}
\arguments{
\item{x}{data \code{\link[dplyr]{tbl_df}} or \code{\link{tbl_svy}} object.}
}
\description{
These functions do not perform non-standard evaluation, and
so are useful when programming against \code{tbl} objects.
\code{ungroup} is a convenient inline way of removing existing
grouping.
}
\seealso{
\code{\link[dplyr]{groups}} for information.
}

