% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interact}
\alias{interact}
\alias{is.interaction}
\title{Create interaction terms to group by when summarizing}
\usage{
interact(...)

is.interaction(x)
}
\arguments{
\item{...}{variables to group by. All tbls accept variable names, some will
also accept functions of variables.}
}
\value{
A vector of type \code{srvyr_interaction}, which is expected to be
invisible to the user.
}
\description{
Allows multiple grouping by multiple variables as if they were a single
variable, which allows calculating proportions that sum to 100% across
more than a single grouping variable with \code{survey_mean}.
}
\details{
Behind the scenes, this function creates a special column type that is
split back into the component columns automatically by \code{summarize}.
}
\examples{
data(api, package = "survey")

dstrata <- apistrat \%>\%
  as_survey_design(strata = stype, weights = pw)

# The sum of prop is equal to 100\%
dstrata \%>\%
  group_by(interact(stype, awards)) \%>\%
  summarize(prop = survey_mean())

# But if you didn't interact, the sum of each stype's prop is 100\%
dstrata \%>\%
  group_by(stype, awards) \%>\%
  summarize(prop = survey_mean())
}
