% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.r
\name{mutate_each}
\alias{funs}
\alias{funs_}
\alias{mutate_each}
\alias{mutate_each_}
\alias{summarise_each}
\alias{summarise_each_}
\alias{summarize_each}
\alias{summarize_each_}
\title{Summarise and mutate multiple columns.}
\description{
See \code{\link[dplyr]{mutate_each}} for more details.
}

