% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_map.R, R/reframe.R, R/split_and_nest.R,
%   R/svy_context.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_map}
\alias{group_modify}
\alias{group_walk}
\alias{reframe}
\alias{group_split}
\alias{group_nest}
\alias{nest_by}
\alias{n}
\alias{cur_data}
\alias{cur_group}
\alias{cur_group_id}
\alias{cur_column}
\alias{across}
\alias{c_across}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:across]{across()}}, \code{\link[dplyr:c_across]{c_across()}}, \code{\link[dplyr:cur_column]{cur_column()}}, \code{\link[dplyr:cur_data]{cur_data()}}, \code{\link[dplyr:cur_group]{cur_group()}}, \code{\link[dplyr:cur_group_id]{cur_group_id()}}, \code{\link[dplyr:group_map]{group_map()}}, \code{\link[dplyr:group_modify]{group_modify()}}, \code{\link[dplyr:group_nest]{group_nest()}}, \code{\link[dplyr:group_split]{group_split()}}, \code{\link[dplyr:group_walk]{group_walk()}}, \code{\link[dplyr:n]{n()}}, \code{\link[dplyr:nest_by]{nest_by()}}, \code{\link[dplyr:reframe]{reframe()}}}
}}

