% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-rlang.R
\name{rlang-tidyeval}
\alias{rlang-tidyeval}
\alias{quo}
\alias{quos}
\alias{enquo}
\alias{quo_name}
\alias{sym}
\alias{syms}
\alias{expr}
\alias{exprs}
\alias{enexpr}
\title{Tidy eval helpers from rlang}
\description{
These functions are used for tidy evaluation, introduced in rlang
and dplyr 0.7. See the programming vignette from dplyr or
<http://rlang.tidyverse.org/articles/tidy-evaluation.html>
for more details.
}
\keyword{internal}
