% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{srvyr_interaction}
\alias{srvyr_interaction}
\title{srvyr interaction column}
\description{
\code{srvyr_interaction} columns help calculate proportions of the interaction of 2
or more variables. They are created by \code{\link{interact}}, generally
used as grouping variables in \code{\link{group_by}} and then automatically split
apart by \code{\link{summarise}}.
}
