% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarise.r
\name{summarise}
\alias{summarise}
\alias{summarise_}
\alias{summarize}
\alias{summarize_}
\alias{survey_mean}
\alias{survey_median}
\alias{survey_quantile}
\alias{survey_ratio}
\alias{survey_total}
\alias{unweighted}
\title{Summarise multiple values to a single value.}
\usage{
summarise(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_svy} object}

\item{...}{Name-value paris of summary functions}
}
\description{
Summarise multiple values to a single value.
}
\details{
\describe{
Summarise for \code{tbl_svy} objects accepts several specialized functions:
\item{\code{survey_mean(..., na.rm = FALSE, vartype = c("se", "ci", "var"))}}{
   Calculate the survey mean of the entire population or by \code{groups}.}
\item{\code{survey_total(..., na.rm = FALSE, vartype = c("se", "ci", "var"))}}{
   Calculate the survey total of the entire population or by \code{groups}.}
 \item{\code{survey_ratio(..., na.rm = FALSE, vartype = c("se", "ci", "var"))}}{
   Calculate the ratio of 2 variables in the entire population or by \code{groups}.}
\item{\code{survey_quantile(..., quantiles, na.rm = FALSE, vartype = c("se", "ci"))}}{
   Calculate quantiles in the entire population or by \code{groups}.}
 \item{\code{survey_median(..., na.rm = FALSE, vartype = c("se", "ci"))}}{
   Calculate the median in the entire population or by \code{groups}.}
 \item{\code{unweighted(...)}}{
   Calculate an unweighted estimate.}
}
}
\examples{
library(survey)
data(api)

dstrata <- apistrat \%>\%
  design_survey(strata = stype, weights = pw)

dstrata \%>\%
  summarise(api99 = survey_mean(api99),
            api00 = survey_mean(api00),
            api_diff = survey_mean(api00 - api99))

dstrata_grp <- dstrata \%>\%
  group_by(stype)

dstrata_grp \%>\%
  summarise(api99 = survey_mean(api99),
            api00 = survey_mean(api00),
            api_diff = survey_mean(api00 - api99))
}

