% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_package_functions.R
\name{svychisq}
\alias{svychisq}
\title{Chisquared tests of association for survey data.}
\arguments{
\item{formula}{See details in \code{\link[survey]{svychisq}}}

\item{design}{See details in \code{\link[survey]{svychisq}}}

\item{na.rm}{See details in \code{\link[survey]{svychisq}}}

\item{...}{See details in \code{\link[survey]{svychisq}}}
}
\description{
Chisquared tests of association for survey data.
}
