% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svychisq_posthoc_test.R
\name{svychisq_posthoc_test}
\alias{svychisq_posthoc_test}
\title{Perform post hoc analysis based on residuals of Pearson's Chi-squared Test for weighted survey data}
\usage{
svychisq_posthoc_test(
  design,
  independent_var,
  dependent_var,
  method = "bonferroni",
  round = 4,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{design}{survey design object from}

\item{independent_var}{independent var for chisq test}

\item{dependent_var}{dependent_var for chisq test}

\item{method}{The p adjustment method to be used. This is passed on to the p.adjust function.}

\item{round}{Number of digits to round the p.value to. Defaults to 4.}

\item{alpha}{The alpha value to the compare the p values to. Defaults to 0.05.}

\item{...}{Additional arguments passed on to the chisq.test function.}
}
\value{
A table with the adjusted p value for each x y combination.
}
\description{
Perform post hoc analysis based on residuals of Pearson's Chi-squared Test for weighted survey data
}
\examples{
# Data from Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender = c("F", "M"),
                   party = c("Democrat","Independent", "Republican"))

# Pass data matrix to chisq.posthoc.test function
chisq.posthoc.test(M, alpha = 0.05)
}
\references{
Agresti, A. (2007). \emph{An Introduction to Categorical Data Analysis}, 2nd
ed. New York: John Wiley & Sons. Page 38.

Beasley, T. M., & Schumacker, R. E. (1995). Multiple Regression Approach
to Analyzing Contingency Tables: Post Hoc and Planned Comparison
Procedures. \emph{The Journal of Experimental Education}, 64(1), 79--93.
https://github.com/ebbertd/chisq.posthoc.test/blob/master/R/chisq_posthoc_test.R
}
