% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{uninteract}
\alias{uninteract}
\alias{uninteract.srvyr_interaction}
\alias{uninteract.data.frame}
\alias{is.interaction}
\title{Break interaction vectors back into component columns}
\usage{
uninteract(x)

\method{uninteract}{srvyr_interaction}(x)

\method{uninteract}{data.frame}(x)

is.interaction(x)
}
\arguments{
\item{x}{Either a \code{srvyr_interaction} column or a \code{data.frame}}
}
\value{
A \code{data.frame}
}
\description{
This function will not generally be needed by users because \code{summarise}
automatically un-interacts interaction columns for you.
}
