% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_nest.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_split}
\alias{group_nest}
\alias{nest_by}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_nest}}, \code{\link[dplyr]{group_split}}, \code{\link[dplyr]{nest_by}}}
}}

