% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarise.r
\name{summarise}
\alias{summarise}
\alias{summarise_}
\alias{summarize}
\alias{summarize_}
\alias{survey_mean}
\alias{survey_total}
\title{Summarise multiple values to a single value.}
\usage{
summarise(.data, ...)
}
\arguments{
\item{.svy}{A \code{tbl_svy} object}

\item{...}{Name-value paris of summary functions}
}
\description{
Summarise multiple values to a single value.
}
\details{
\describe{
Summarise for \code{tbl_svy} objects accepts only 1 function:
\item{\code{survey_mean(..., vartype = c("se", "ci", "var")}}{
   Calculate the survey mean of the entire population or by \code{groups}.}
}
}
\examples{
library(survey)
data(api)

dstrata <- apistrat \%>\%
  design_survey(strata = stype, weights = pw)

dstrata \%>\%
  summarise(api99 = survey_mean(api99),
            api00 = survey_mean(api00),
            api_diff = survey_mean(api00 - api99))

dstrata_grp <- dstrata \%>\%
  group_by(stype)

dstrata_grp \%>\%
  summarise(api99 = survey_mean(api99),
            api00 = survey_mean(api00),
            api_diff = survey_mean(api00 - api99))
}

