% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.r
\name{summarise_all}
\alias{summarise_all}
\alias{summarize_all}
\alias{summarise_if}
\alias{summarize_if}
\alias{summarise_at}
\alias{summarize_at}
\alias{mutate_all}
\alias{mutate_if}
\alias{mutate_at}
\alias{mutate_each}
\alias{mutate_each_}
\alias{summarise_each}
\alias{summarise_each_}
\alias{summarize_each}
\alias{summarize_each_}
\alias{funs}
\alias{funs_}
\title{Summarise and mutate multiple columns.}
\description{
See \code{\link[dplyr]{summarize_all}} for more details. *_each functions will be depracated
in favor of *_all/*_if/*_at functions, but the latter do not yet work on database backed tables.
}
