% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_statistics.r
\name{survey_total}
\alias{survey_total}
\title{Calculate the total and its variation using survey methods}
\usage{
survey_total(x = NULL, na.rm = FALSE, vartype = c("se", "ci", "var",
  "cv"), level = 0.95, ...)
}
\arguments{
\item{x}{A variable or expression, or empty}

\item{na.rm}{A logical value to indicate whether missing values should be dropped}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number or vector of numbers indicating the confidence level}

\item{...}{Ignored}
}
\description{
Calculate totals from complex survey data. A wrapper
around \code{\link[survey]{svytotal}}. \code{survey_total} should always be
called from \code{\link{summarise}}.
}
\examples{
library(survey)
data(api)

dstrata <- apistrat \%>\%
  as_survey_design(strata = stype, weights = pw)

dstrata \%>\%
  summarise(enroll = survey_total(enroll),
            tot_meals = survey_total(enroll * meals / 100, vartype = c("ci", "cv")))

dstrata \%>\%
  group_by(awards) \%>\%
  summarise(api00 = survey_total(enroll))

# Leave x empty to calculate the total in each group
dstrata \%>\%
  group_by(awards) \%>\%
  summarise(pct = survey_total())

# level takes a vector for multiple levels of confidence intervals
dstrata \%>\%
  summarise(enroll = survey_total(enroll, vartype = "ci", level = c(0.95, 0.65)))

}

